pyproject.toml
[project]
name = "vast"
version = "0.1.0"
description = ""
authors = [
    {name = "Mengqing Hu",email = "mengqinghu688@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "numpy (>=2.2.0,<3.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "openai-whisper (>=20250625,<20250626)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "pytube (>=15.0.0,<16.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "notebook (>=7.4.7,<8.0.0)",
    "ipykernel (>=7.0.1,<8.0.0)",
    "yt-dlp (>=2025.10.22,<2026.0.0)",
    "torch (>=2.9.0,<3.0.0)",
    "torchvision (>=0.24.0,<0.25.0)",
    "torchaudio (>=2.9.0,<3.0.0)",
    "python-box (>=7.3.2,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
